Fix/dproxy (#56)
This finally fixes dproxy for tools like accessibility inspector and instruments. They have this weird behavior that they disable SSL after stopping a session, just to re-enable it on the same connection. 
Since golang does not really support such a scenario (and why would they) the code is a bunch of hacks to make it work. 
The missing piece was setting the WriteTimeout on the net.Conn to something that does not immediately timeout after sending the closeWrite() command. 
Accessibility Inspector still crashes for some reason, but this is better than before.