Add option to pretty-print JSON command output. Add trailing newlines to command output.
This commit adds a the --pretty option to format any JSON output with a 4-space indent using golang's json.MarshalIndent, making certain output easier to read when working manually on the command-line.

While updating a few printf calls, I also noticed a few commands that I believe were unintentionally omitting a trailing newline, so that after running those commands the user's command prompt would be at the end of the output of the previous command.  In principle this *could* break scripted commands that assume no newline? Hard to imagine...