Add iOS11+ version check to "assistivetouch" command, with --force override to attempt anyway.
Having tested on a grand total of 1 iOS10.x device (an iPhone 6s plus), we speculate that the technique used to enable/disable assistive touch does not work on devices older than iOS 11.  This commit adds a fail-fast check for versions older than iOS 11.

However, since only 1 device has been tested, (and older iPad OS has not been tested at all), I have also added a "--force" override so that the end-user can still give it a try.

The "--force" option also forces SetValue() to be called even if it would not cause an apparent change. (By default, if assistive touch is already disabled, and the user asks to disable it, nothing happens).

"--force" is mostly meant as a way to continue testing older device functionality without having to have a custom debug build of go-ios.  If we can definitively prove that iOS12 is the minimum, the --force option can be removed.