Add ability to enable/disable AssistiveTouch soft "home" button. iOS 11+ (#148)
* Add ability to enable/disable AssistiveTouch soft "home" button. iOS 11+

Using the same technique already used for querying and setting the device language and locale, go-ios can now enable and disable AssistiveTouch using the the magic boolean com.apple.Accessibility.AssistiveTouchEnabledByiTunes

This commit slightly modifies the existing SetValueForDomain() and related methods and structs to take an interface{} value (rather than a string) in order to pass a raw bool to lockdown.

This commit also replaces a few inconsistently used tabs in the verbose command-line "help" output.

Thank you to the team at alibaba/taobao-iphone-device (tidevice) for sharing the magic word to make this happen.

Validated on iPhone 5s, 6, 6s, 6s plus, 7, 8, se 1st gen, se 2nd gen, and 11 pro max.  iOS 11.4.1, 12.1, 12.5.5, 13.4, 13.6, 14.0.1, 14.7, 15.0, 15.3, and 16.0.1 beta.

Failed on an iPhone 6+ running iOS 10.2.1 (Exits with error including hint: “Is this device running iOS 11+?”)

* Add iOS11+ version check to "assistivetouch" command, with --force override to attempt anyway.

Having tested on a grand total of 1 iOS10.x device (an iPhone 6s plus), we speculate that the technique used to enable/disable assistive touch does not work on devices older than iOS 11.  This commit adds a fail-fast check for versions older than iOS 11.

However, since only 1 device has been tested, (and older iPad OS has not been tested at all), I have also added a "--force" override so that the end-user can still give it a try.

The "--force" option also forces SetValue() to be called even if it would not cause an apparent change. (By default, if assistive touch is already disabled, and the user asks to disable it, nothing happens).

"--force" is mostly meant as a way to continue testing older device functionality without having to have a custom debug build of go-ios.  If we can definitively prove that iOS12 is the minimum, the --force option can be removed.

Co-authored-by: danielpaulus <daniel.paulus@icloud.com>