fail if more than one device is connected, add panic handlers to move back socket (#206)
I finally got that in golang to properly handle panics, you have to add the panic handler in every goroutine. I added that so dproxy will now always fix and move back the broken socket even if it fails with a panic. 
Also, dproxy currently does not work with multiple devices. I made it fail with an error in case more than one device is connected.