Add mjpeg server to screenshotr (#213)
start a simple mjpeg server by running
`ios screenshot --stream` on 0.0.0.0:3333. If you provide the port argument like so: 
`ios screenshot --stream --port=8082` the server will run on your specified port. Then watch a slow stream in your browser. 
This is only a fun feature or to use in debugging if all else fails. Screenshotr is too slow to be really useable on Retina/High Res devices because it sends huge PNG files. Those reach memory bandwith limits on the device. Also it is too unstable to be used in production scenarios. For useable mjpeg streams, use WebDriverAgent's built in server. 
On older devices, this works quite well though :-D