Get all XCUITest logs (#398)
Starting with iOS 17 XCUITest logs are sent over two channels. There are debug logs that come over the DTX connection (like they did in the past), but the non-debug messages are sent over a different connection to the com.apple.coredevice.openstdiosocket service.

We now open a connection to the openstdiosocket-service before we launch the test runner, and we tell the appservice to which socket it should connect stdio. And after that we pipe the data we get from the stdio-socket to the logoutput.