Add missing proxydispatcher cases + fix productModuleName for iOS <17 (#547)
The method switch in proxydispatcher::Dispatch was missing a few cases, which led to missing XCUITest results when using go-ios on iOS <17 devices.

Additionally, the productModuleName field in the XCTestConfiguration appeared to be wrong (also for iOS <17 devices), leading to the same issue of missing test results. It has been changed to match the iOS 17 config, which seems to work better for the test cases I've tried.