Refactor tests to be DAMP instead of DRY
The helper method to load the different test targets, required the reader to make a lot of mental
jumps to understand which test target was actually used in a test. Tests are not production code and should
be as simple to follow along as possible. Now, when a reader wants to know which target is used, they know that
they should look at the file that is located at /testdata/format_version2... and they know that they need to look at the
first or second entry of the TestConfiguration array.
This is important when the reader wants to understand the relation between the xctestrun file and the code.