Persistent go-ios agent after the initial go-ios terminates. (#560)
When go-ios is started in agent mode, the agent terminates when the parent go-ios process terminates. If you run go-ios in multiple terminal for multiple device test automation the this can disrupt other terminal sessions which may be using the first started tunnel.

With this PR we use SysProcAttr for *nix and Window platforms.

For *nix platforms setting Setsid: true starts agent in an independent new *nix session. And for windows as this setting is not available by OS nature we can use CreationFlags: syscall.CREATE_NEW_PROCESS_GROUP to start the process in a new process group.