list icons using springboard service (#609)
Apple Configurator 2 has this feature of modifying the layout of the home screen on the device. One part of this feature is showing how the layout looks like at the moment, and this is what we provide with this change.

We get the list of icons, and their type, for each screen of the home screen