Refactor AFC (#627)
The current implementation has a lot of repeated code. Every method that communicates with the afc service on the device has it's own logic for building the package and maintaining the package identifiers.

This is now done in a single place that abstracts away of this logic that was repeated over and over before.

Also, file I/O works now similar to "regular" filesystems. When we open a file using the afc service, we get back a io.ReadWriteCloser, which allows us to use functions like io.Copy on those files.

---------

Co-authored-by: Andreas Lüdeke <311702+aluedeke@users.noreply.github.com>
Co-authored-by: Anton Sokolchenko <wsevendays@gmail.com>